+++ src/wizelit_agent_wrapper/__init__.py
@@ -2,6 +2,8 @@
 
 __version__ = "0.1.0"
 
-# Import your main modules here
-# from .your_module import your_function
+# Import main functions
+from .utils import greet
+
+__all__ = ["greet", "__version__"]